### 2024-11-15 #7485 @crvernon
:wave: @NiklasMelton - while I am getting you a topic editor assigned, please reduce the length of your paper to around 1000 words.  Also, please check to ensure that you have DOI for all citations where available.  Thanks!
