### 2024-11-15 #7485 @vissarion
Hi Chris, I am already behind in https://github.com/openjournals/joss-reviews/issues/7426 so if this is not urgent you can assign it to me.
