+++ references.bib
@@ -1,12 +1,15 @@
 @inproceedings{bezdek2002vat,
-  title={VAT: A tool for visual assessment of (cluster) tendency},
-  author={Bezdek, James C and Hathaway, Richard J},
-  booktitle={Proceedings of the 2002 International Joint Conference on Neural Networks. IJCNN'02 (Cat. No. 02CH37290)},
-  volume={3},
-  pages={2225--2230},
+  author={Bezdek, J.C. and Hathaway, R.J.},
+  booktitle={Proceedings of the 2002 International Joint Conference on Neural Networks. IJCNN'02 (Cat. No.02CH37290)},
+  title={VAT: a tool for visual assessment of (cluster) tendency},
   year={2002},
-  organization={IEEE}
-}
+  volume={3},
+  number={},
+  pages={2225-2230 vol.3},
+  keywords={Computer science;Data analysis;Clustering methods;Clustering algorithms;Symmetric matrices;Pattern recognition;Mathematical model;Digital images;Pixel;Mathematics},
+  doi={10.1109/IJCNN.2002.1007487}
+ }
+
 @article{da2022icvi,
   title={iCVI-ARTMAP: using incremental cluster validity indices and adaptive resonance theory reset mechanism to accelerate validation and achieve multiprototype unsupervised representations},
   author={da Silva, Leonardo Enzo Brito and Rayapati, Nagasharath and Wunsch, Donald C},
@@ -15,7 +18,9 @@
   number={12},
   pages={9757--9770},
   year={2022},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TNNLS.2022.3160381}
+
 }
 @article{carpenter1987massively,
   title={A massively parallel architecture for a self-organizing neural pattern recognition machine},
@@ -25,7 +30,8 @@
   number={1},
   pages={54--115},
   year={1987},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/S0734-189X(87)80014-2}
 }
 @article{carpenter1987art,
   title={ART 2: Self-organization of stable category recognition codes for analog input patterns},
@@ -35,7 +41,8 @@
   number={23},
   pages={4919--4930},
   year={1987},
-  publisher={Optica Publishing Group}
+  publisher={Optica Publishing Group},
+  doi={10.1364/AO.26.004919}
 }
 @article{carpenter1991art,
   title={ART 2-A: An adaptive resonance algorithm for rapid category learning and recognition},
@@ -55,7 +62,8 @@
   number={6},
   pages={1628--1644},
   year={2007},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1016/0893-6080(91)90045-7}
 }
 @inproceedings{anagnostopoulos2001a,
   title={Ellipsoid ART and ARTMAP for incremental clustering and classification},
@@ -64,7 +72,8 @@
   volume={2},
   pages={1221--1226},
   year={2001},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/IJCNN.2001.939535}
 }
 @inproceedings{anagnostopoulos2001b,
   title={Ellipsoid ART and ARTMAP for incremental unsupervised and supervised learning},
@@ -73,7 +82,8 @@
   volume={4390},
   pages={293--304},
   year={2001},
-  organization={SPIE}
+  organization={SPIE},
+  doi = {10.1117/12.421180}
 }
 @article{carpenter1991fuzzy,
   title={Fuzzy ART: Fast stable learning and categorization of analog patterns by an adaptive resonance system},
@@ -83,7 +93,8 @@
   number={6},
   pages={759--771},
   year={1991},
[…]
