+++ references.bib
@@ -270,7 +270,7 @@
  pages         = {357--362},
  doi           = {10.1038/s41586-020-2649-2},
  publisher     = {Springer Science and Business Media {LLC}},
- url           = {https://doi.org/10.1038/s41586-020-2649-2}
+ url           = {https://doi.org/10.1038/s41586-020-2649-2},
  doi           = {10.1038/s41586-020-2649-2}
 }
 @article{da2019survey,
