+++ references.bib
@@ -314,7 +314,7 @@
   volume={87 1},
   pages={
           1-51
-        }
+        },
   doi={10.1007/978-94-009-7758-7_1}
 }
 @article{grossberg2013adaptive,
