### 2024-11-20 7fa3727
update hull-art for 3D
