+++ examples/demo_convex_hull_art.py
@@ -1,50 +1,17 @@
-from sklearn.datasets import make_blobs, make_moons, make_swiss_roll
+from sklearn.datasets import make_moons, make_blobs
 import matplotlib.pyplot as plt
 import numpy as np
 
 from artlib.experimental.HullART import HullART
-from artlib import VAT
-from matplotlib.path import Path
 
-
-
-def cluster_blobs():
-    data, target = make_blobs(
-        n_samples=150,
-        centers=1,
-        cluster_std=0.50,
-        random_state=0,
-        shuffle=False,
-    )
-
-    print("Data has shape:", data.shape)
-
-    params = {"rho": 0.1, "alpha": 1e-3, "alpha_hull": 2.0, "min_lambda": 1e-10}
-    cls = HullART(**params)
-
-    X = cls.prepare_data(data)
-    print("Prepared data has shape:", X.shape)
-
-    y = cls.fit_predict(X)
-
-    print(f"{cls.n_clusters} clusters found")
-
-    cls.visualize(X, y)
-    plt.show()
 
 def cluster_moons():
     data, target = make_moons(n_samples=1000, noise=0.1, random_state=10,
                               shuffle=False)
-    # vat, idx = VAT(data)
-    # plt.figure()
-    # plt.imshow(vat)
-    #
-    # data = data[idx,:]
-    # target = target[idx]
     print("Data has shape:", data.shape)
 
     params = {"rho": 0.2, "alpha": 1e-20, "alpha_hull": 8., "min_lambda": 1e-5,
-              "max_lambda": 0.2}
+              "max_lambda": 0.4}
     cls = HullART(**params)
 
     X = cls.prepare_data(data)
@@ -55,12 +22,39 @@ def cluster_moons():
     y = cls.labels_
     print(np.unique(y))
 
-    print(f"{cls.n_clusters} clusters found")
-
     cls.visualize(X, y)
     plt.show()
 
 
+def cluster_3d_data():
+    # Generate 3D dataset
+    data, target = make_blobs(n_samples=1000, centers=4, n_features=3,
+                              cluster_std=1.0, random_state=9, shuffle=False)
+    print("Data has shape:", data.shape)
+
+    # Parameters for HullART
+    params = {"rho": 0.0, "alpha": 1e-20, "alpha_hull": 0., "min_lambda": 1e-1,
+              "max_lambda": 0.4}
+    cls = HullART(**params)
+
+    # Prepare data
+    X = cls.prepare_data(data)
+    print("Prepared data has shape:", X.shape)
+
+    # Fit the model
+    cls = cls.fit(X, verbose=True)
+    y = cls.labels_
+    print(np.unique(y))
+
+
+    # Visualization for 3D data
+    fig = plt.figure(figsize=(10, 7))
+    ax = fig.add_subplot(111, projection='3d')
+    cls.visualize(X, y, ax=ax)
+    plt.show()
+    plt.show()
+
+
 if __name__ == "__main__":
-    # cluster_blobs()
-    cluster_moons()
+    # cluster_moons()
+    cluster_3d_data()
