### 2024-11-25 #7485 @mstimberg
👋 @NiklasMelton I will be the editor for your submission, looking forward to working with you.

The first step is to look for reviewers. I will do this mostly over email, so don't be surprised if you won't see much activity here. If you have any suggestions for potential reviewers, then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
