### 2024-11-27 0b52c4f
hull art tests
