### 2024-11-27 8d0d9a1
add scripts
