+++ hull_art_tests.py
@@ -1,0 +1,524 @@
+from artlib import FuzzyART, SimpleARTMAP, HypersphereART, VAT
+from artlib.experimental.HullART import HullART
+import numpy as np
+from ucimlrepo import fetch_ucirepo
+import pandas as pd
+from sklearn.model_selection import StratifiedKFold
+from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
+import time
+import pickle
+import warnings
+from tqdm import tqdm
+import os
+
+os.environ["TF_ENABLE_ONEDNN_OPTS"] = "0"
+import tensorflow as tf
+import concurrent.futures
+
+warnings.filterwarnings("ignore")
+
+datasets = [
+    # (15, "breast cancer (original)"),
+    (936, "NPHA"),
+    (759, "glioma"),
+    (17, "breast cancer (diagnostic)"),
+    (14, "breast cancer (yugoslavia)"),
+    (16, "breast cancer (prognostic)"),
+    (451, "breast cancer (coimbra)"),
+    (45, "heart disease"),
+    (519, "heart failure"),
+    (62, "lung cancer"),
+    # (46, "hepatitis"),
+    (174, "parkinsons"),
+    (145, "statlog"),
+    (33, "dermatology"),
+    (30, "contraceptive choice"),
+    (225, "ILPD"),
+    # (571, "HCV"),
+    # (43, "habermans"),
+    (863, "maternal health"),
+    (915, "differntiated thyroid cancer"),
+    (890, "AIDS clinical trials 175"),
+    (878, "cirrhosis"),
+    (244, "fertility"),
+    (83, "primary tumor"),
+]
+
+art_models = [
+    (FuzzyART, {"rho": 0.0, "alpha": 1e-5, "beta": 1.0}),
+    (HypersphereART, {"rho": 0.0, "alpha": 1e-5, "beta": 1.0, "r_hat": None}),
+    (
+        HullART,
+        {
+            "rho": 0.0,
+            "alpha": 1e-5,
+            "alpha_hull": 0.0,
+            "min_lambda": 1e-5,
+            "max_lambda": 0.2,
+        },
+    ),
+]
+
+MT_METHODS = ["MT1", "MT+", "MT-", "MT0", "MT~"]
+EPSILONS = [1e-1, 1e-2, 1e-3, 1e-4, 1e-5, np.inf]
+
+
+def verify_numeric(arr):
+    return np.issubdtype(arr.dtype, np.number)
+
+
+def load_mnist_as_numpy():
+    # Load the MNIST dataset
+    (x_train, y_train), (x_test, y_test) = tf.keras.datasets.mnist.load_data()
+
+    # Flatten the images from 28x28 into 784 (28*28) vectors
+    x_train_flat = x_train.reshape(x_train.shape[0], -1)
+    x_test_flat = x_test.reshape(x_test.shape[0], -1)
+
+    # Normalize the pixel values to the range [0, 1]
+    x_train_flat = x_train_flat.astype("float32") / 255
+    x_test_flat = x_test_flat.astype("float32") / 255
+
+    # # Combine train and test data into one numpy array
+    # X = np.concatenate([x_train_flat, x_test_flat])
+    # y = np.concatenate([y_train, y_test]).reshape((-1, 1))
+
+    # return X, y
+    y = y_test.reshape((-1, 1))
+
+    return x_test_flat, y
+
+
+def data_loader():
+    for id, name in datasets:
+        try:
+            if name == "MNIST":
+                X, y = load_mnist_as_numpy()
+                yield name, X, y
+            else:
+                if os.path.exists(f"data/uci_dataset_{name}_{id}.pickle"):
+                    data = pickle.load(
+                        (open(f"data/uci_dataset_{name}_{id}.pickle", "rb"))
+                    )
[…]
