### 2024-11-30 #7485 @NiklasMelton
> 👋 @NiklasMelton I will be the editor for your submission, looking forward to working with you.
>
> The first step is to look for reviewers. I will do this mostly over email, so don't be surprised if you won't see much activity here. If you have any suggestions for potential reviewers, then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

I know [@AP6YC](https://reviewers.joss.theoj.org/reviewers/2389) has familiarity with this area but not sure if he has time.

Otherwise, I think anyone appearing on this search would be suitable:
https://reviewers.joss.theoj.org/reviewers/search?language=26&area_id=5&keywords=clustering&name=&commit=Search
