### 2024-12-07 #7485 @AP6YC
@NiklasMelton Thank you for recommending me as a reviewer! @mstimberg As I mentioned via email, I would be more than happy to review this submission, but I must disclose a potential COI per the JOSS COI policy: Nik is a colleague of mine in the same laboratory and university department where I am now concluding my Ph.D. and where I am beginning a faculty position in January 1st. If @mstimberg judges that this does not preclude me from reviewing the submission, I would be delighted to do so.
