### 2024-12-09 #7485 @mstimberg
Dear @AP6YC, thanks for getting back to us and for your willingness to help out. Unfortunately, our policies state that "In addition, your recent (past year) association with the same organization of a submitter is a COI, for example, being employed at the same institution." which applies to your situation – apologies for not checking your affiliations earlier…
This means we won't be able to add you as a reviewer, I am afraid. Please don't hesitate to suggest other potential reviewers (here – without tagging them – or over email) in case you can think of someone. Thanks!
