### 2025-01-09 8b495e8
improve error message and add helper functions
