### 2025-01-09 69a0d0e
default is_fitted
