### 2025-01-09 2b83265
fix some broken unit tests
