+++ unit_tests/test_FuzzyART.py
@@ -202,10 +202,11 @@ def test_find_data_bounds(art_model):
 def test_prepare_data(art_model):
     # Test prepare_data with valid data
     X = np.array([[0.0, 0.5], [0.2, 1.0]])
+    X_cc = compliment_code(X)
     art_model.d_min_ = np.array([0.0, 0.0])
     art_model.d_max_ = np.array([1.0, 1.0])
     normalized_X = art_model.prepare_data(X)
-    np.testing.assert_array_almost_equal(normalized_X, X)  # Already normalized
+    np.testing.assert_array_almost_equal(normalized_X, X_cc)  # Already normalized
 
     # Test prepare_data with data requiring normalization
     X = np.array([[1.0, 10.0], [5.0, 20.0]])
@@ -213,4 +214,5 @@ def test_prepare_data(art_model):
     art_model.d_max_ = np.array([5.0, 20.0])
     normalized_X = art_model.prepare_data(X)
     expected_normalized_X = np.array([[0.0, 0.0, 1.0, 1.0], [1.0, 1.0, 0.0, 0.0]])
-    np.testing.assert_array_almost_equal(normalized_X, expected_normalized_X)
+    expected_normalized_X_cc = compliment_code(expected_normalized_X)
+    np.testing.assert_array_almost_equal(normalized_X, expected_normalized_X_cc)
