### 2025-01-09 0715e33
fix some broken unit tests
