### 2025-01-09 58c368a
update docs and add clip method
