+++ artlib/supervised/ARTMAP.py
@@ -257,13 +257,15 @@ class ARTMAP(SimpleARTMAP):
         )
         return self
 
-    def predict(self, X: np.ndarray) -> np.ndarray:
+    def predict(self, X: np.ndarray, clip: bool = False) -> np.ndarray:
         """Predict the labels for the given data.
 
         Parameters
         ----------
         X : np.ndarray
             Data set A (independent channel).
+        clip : bool
+            clip the input values to be between the previously seen data limits
 
         Returns
         -------
@@ -272,15 +274,19 @@ class ARTMAP(SimpleARTMAP):
 
         """
         check_is_fitted(self)
-        return super(ARTMAP, self).predict(X)
+        return super(ARTMAP, self).predict(X, clip)
 
-    def predict_ab(self, X: np.ndarray) -> tuple[np.ndarray, np.ndarray]:
+    def predict_ab(
+        self, X: np.ndarray, clip: bool = False
+    ) -> tuple[np.ndarray, np.ndarray]:
         """Predict both A-side and B-side labels for the given data.
 
         Parameters
         ----------
         X : np.ndarray
             Data set A (independent channel).
+        clip : bool
+            clip the input values to be between the previously seen data limits
 
         Returns
         -------
@@ -289,7 +295,7 @@ class ARTMAP(SimpleARTMAP):
 
         """
         check_is_fitted(self)
-        return super(ARTMAP, self).predict_ab(X)
+        return super(ARTMAP, self).predict_ab(X, clip)
 
     def predict_regression(self, X: np.ndarray) -> np.ndarray:
         """
