### 2025-01-09 c6c0cfa
update docs and add clip method
