### 2025-01-10 89391c9
update readme examples with mnist and normalization info
