### 2025-01-10 6442d95
update readme examples with mnist and normalization info
