### 2025-02-10 #7764 @editorialbot
[REVIEW]: Adaptive Resonance Lib: A Python package for Adaptive Resonance Theory (ART) models
**Submitting author:** @NiklasMelton (<a href="http://orcid.org/0000-0001-9625-7086">Niklas Melton</a>)
**Repository:** https://github.com/NiklasMelton/AdaptiveResonanceLib
**Branch with paper.md** (empty if default branch): develop
**Version:** v0.1.3
