### 2025-02-10 #7764 @mstimberg
👋🏼 @NiklasMelton @TahiriNadia, @chenxinye this is the review thread for the paper. All of our communications will happen here from now on.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/7764#issue-2842721452).
