### 2025-02-18 0dc5a0c
improve verbose fit for nested tqdm
