### 2025-02-18 eb40dad
handle nans in normalization
