### 2025-02-18 d6ff4ca
optimize art1
