### 2025-02-18 b8ab4d0
add numba optimizations to ART1
