### 2025-02-18 3b32531
optimize utils and FuzzyART with numba
