### 2025-02-18 78afe7c
scaling factor back to variable
