### 2025-02-18 4111c26
correct bugs with weight types
