### 2025-02-18 4766da6
update art1 unit tests
