### 2025-02-18 c04ea48
update art1 unit tests
