### 2025-02-18 f98705d
revert utils as the numba optimization there doesnt help
