### 2025-02-19 125bf7d
revert numba change for ART1
