### 2025-02-19 1f41ecd
revert numba change for ART1
