### 2025-02-24 58d1502
add binary Fuzzy ART
