+++ artlib/elementary/BinaryFuzzyART.py
@@ -1,0 +1,229 @@
+"""Fuzzy ART :cite:`carpenter1991fuzzy`."""
+# Carpenter, G. A., Grossberg, S., & Rosen, D. B. (1991c).
+# Fuzzy ART: Fast stable learning and categorization of analog patterns by an
+# adaptive resonance system.
+# Neural Networks, 4, 759 – 771. doi:10.1016/0893-6080(91)90056-B.
+
+from artlib.elementary.FuzzyART import FuzzyART
+from typing import Optional, Callable, Literal, Tuple, Dict
+import warnings
+import numpy as np
+from numba import njit
+
+
+@njit
+def _category_choice_binary(
+    i: np.ndarray, w: np.ndarray, alpha: float, pre_MT: bool, rho_w1: int
+) -> Tuple[float, int]:
+    """Optimized category choice for binary data using count_nonzero."""
+    w1 = np.count_nonzero(i & w)
+    if not pre_MT or w1 >= rho_w1:
+        return w1 / (alpha + np.count_nonzero(w)), w1
+    else:
+        return np.nan, w1
+
+
+@njit
+def _match_criterion_binary(i: np.ndarray, w: np.ndarray, dim_original: float) -> float:
+    """Optimized match criterion for binary data using count_nonzero."""
+    return np.count_nonzero(i & w) / dim_original
+
+
+@njit
+def _update_binary(i: np.ndarray, w: np.ndarray) -> np.ndarray:
+    """Optimized update rule for binary data."""
+    return i & w  # Using bitwise AND for binary updates
+
+
+class BinaryFuzzyART(FuzzyART):
+    """Fuzzy ART optimized for binary input data."""
+
+    def __init__(self, rho: float, alpha: float):
+        """Initialize the Fuzzy ARTb model.
+
+        Parameters
+        ----------
+        rho : float
+            Vigilance parameter.
+        alpha : float
+            Choice parameter.
+
+        """
+        super().__init__(rho, alpha, beta=1.0)
+
+    @staticmethod
+    def validate_params(params: dict):
+        """Validate clustering parameters.
+
+        Parameters
+        ----------
+        params : dict
+            Dictionary containing parameters for the algorithm.
+
+        """
+        assert "rho" in params
+        assert "alpha" in params
+        assert 1.0 >= params["rho"] >= 0.0
+        assert params["alpha"] >= 0.0
+        assert isinstance(params["rho"], float)
+        assert isinstance(params["alpha"], float)
+
+    def validate_data(self, X: np.ndarray):
+        """Validate the data prior to clustering.
+
+        Parameters
+        ----------
+        X : np.ndarray
+            Dataset.
+
+        """
+        assert X.shape[1] % 2 == 0, "Data has not been compliment coded"
+        assert X.dtype == np.bool_ or np.issubdtype(
+            X.dtype, np.integer
+        ), "Binary Fuzzy ART only supports binary data"
+        assert ((X == 0) | (X == 1)).all(), "Binary Fuzzy ART only supports binary data"
+        assert np.all(
+            abs(np.sum(X, axis=1) - float(X.shape[1] / 2)) <= 0.01
+        ), "Data has not been compliment coded"
+        self.check_dimensions(X)
+
+    def category_choice(
+        self, i: np.ndarray, w: np.ndarray, params: dict
+    ) -> tuple[float, Optional[dict]]:
+        """Get the activation of the cluster using optimized binary operations."""
+        pre_MT = params["MT"] not in [None, "MT-"]
+        T, w1 = _category_choice_binary(i, w, params["alpha"], pre_MT, params["rho_w1"])
+        return T, {"w1": w1}
+
+    def match_criterion(
+        self,
+        i: np.ndarray,
+        w: np.ndarray,
[…]
