### 2025-02-26 3b12c1d
add cpp optimized Binary Fuzzy ART
