+++ artlib/cpp_optimized/BinaryFuzzyARTMAP.cpp
@@ -1,0 +1,216 @@
+#include <pybind11/pybind11.h>
+#include <pybind11/numpy.h>
+#include <pybind11/stl.h>
+#include <vector>
+#include <unordered_map>
+#include <algorithm>
+#include <cassert>
+#include <cmath>
+
+namespace py = pybind11;
+
+class BinaryFuzzyARTMAP {
+public:
+    struct Cluster {
+        std::vector<int> weight;
+    };
+
+    BinaryFuzzyARTMAP(double rho, double alpha, std::string MT, double epsilon)
+        : base_rho_(rho), alpha_(alpha), MT_(MT), epsilon_(epsilon)
+    {
+        // We do NOT know dim_original_ yet, so we simply set to 0 here.
+        // We'll infer it in fit() from the data.
+        dim_original_ = 0;
+        rho_w1_       = 0;
+    }
+
+    std::tuple<py::array_t<int>, std::vector<py::array_t<int>>, py::array_t<int>> fit(py::array_t<int> X, py::array_t<int> y) {
+        py::buffer_info x_buf = X.request(), y_buf = y.request();
+
+        assert(x_buf.ndim == 2 && "X must be a 2D array");
+        assert(y_buf.ndim == 1 && "y must be a 1D array");
+
+        pybind11::ssize_t num_samples_ssize = x_buf.shape[0];
+        pybind11::ssize_t num_features_ssize = x_buf.shape[1];
+
+        // Ensure values fit in `int`
+        assert(num_samples_ssize <= std::numeric_limits<int>::max());
+        assert(num_features_ssize <= std::numeric_limits<int>::max());
+
+        int num_samples = static_cast<int>(num_samples_ssize);
+        int num_features = static_cast<int>(num_features_ssize);
+
+        dim_original_ = num_features / 2;
+        rho_w1_       = static_cast<int>(base_rho_ * dim_original_);
+
+        auto x_ptr = static_cast<int*>(x_buf.ptr);
+        auto y_ptr = static_cast<int*>(y_buf.ptr);
+
+        std::vector<int> labels(num_samples);
+
+        for (int i = 0; i < num_samples; ++i) {
+            std::vector<int> sample(x_ptr + i * num_features, x_ptr + (i + 1) * num_features);
+            int c_b = y_ptr[i];
+            labels[i] = step_fit(sample, c_b);
+        }
+
+        py::array_t<int> labels_out(labels.size());
+        std::memcpy(labels_out.mutable_data(), labels.data(), labels.size() * sizeof(int));
+
+        std::vector<py::array_t<int>> weight_arrays;
+        weight_arrays.reserve(clusters_.size());
+        for (const auto& cluster : clusters_) {
+            py::array_t<int> arr(cluster.weight.size());
+            std::memcpy(arr.mutable_data(), cluster.weight.data(), cluster.weight.size() * sizeof(int));
+            weight_arrays.push_back(std::move(arr));
+        }
+
+        std::vector<int> cluster_labels(clusters_.size());
+        for (const auto& [cluster_id, label] : cluster_map_) {
+            cluster_labels[cluster_id] = label;
+        }
+        py::array_t<int> cluster_labels_out(cluster_labels.size());
+        std::memcpy(cluster_labels_out.mutable_data(),
+                    cluster_labels.data(),
+                    cluster_labels.size() * sizeof(int));
+
+        return std::make_tuple(labels_out, weight_arrays, cluster_labels_out);
+    }
+
+private:
+    double base_rho_;
+    double rho_;
+    double alpha_;
+    int dim_original_;
+    int rho_w1_;
+    std::string MT_;
+    double epsilon_;
[…]
