### 2025-02-27 a513717
adjust pyproject for c++ build
