+++ pyproject.toml
@@ -51,5 +51,5 @@ sphinxcontrib-bibtex = "^2.0.0"
 
 # Specify build system for C++ extensions
 [build-system]
-requires = ["setuptools", "wheel", "pybind11"]
-build-backend = "setuptools.build_meta"
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
