+++ pyproject.toml
@@ -24,6 +24,7 @@ homepage = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 repository = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 documentation = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 readme = "README.md"
+build = "build_extension.py"
 
 [tool.setuptools.packages.find]
 include = ["artlib*"]
@@ -51,5 +52,5 @@ sphinxcontrib-bibtex = "^2.0.0"
 
 # Specify build system for C++ extensions
 [build-system]
-requires = ["poetry-core>=1.0.0"]
+requires = ["poetry-core>=1.0.0", "setuptools", "pybind11"]
 build-backend = "poetry.core.masonry.api"
