+++ pyproject.toml
@@ -24,6 +24,7 @@ homepage = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 repository = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 documentation = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 readme = "README.md"
+build = "build_extension.py"
 
 [tool.setuptools.packages.find]
 include = ["artlib*"]
