+++ pyproject.toml
@@ -54,6 +54,3 @@ sphinxcontrib-bibtex = "^2.0.0"
 [build-system]
 requires = ["poetry-core>=1.0.0", "setuptools", "pybind11"]
 build-backend = "poetry.core.masonry.api"
-
-[tool.poetry.build]
-script = "build_extension.py"
