+++ pyproject.toml
@@ -1,34 +1,14 @@
-[project]
-name = "artlib"
-version = "0.1.3"
-description = "A Python library for Adaptive Resonance Theory (ART) algorithms."
-authors = [{ name = "Niklas M. Melton", email = "niklasmelton@gmail.com" }]
-license = { text = "MIT" }
-readme = "README.md"
-
-dependencies = [
-    "numpy>=1.19.4",
-    "scipy>=1.5.4",
-    "scikit-learn>=1.0",
-    "matplotlib>=3.3.3",
-    "numba>=0.57.0"
-]
-
 [tool.poetry]
 name = "artlib"
 version = "0.1.3"
 description = "A Python library for Adaptive Resonance Theory (ART) algorithms."
 authors = ["Niklas M. Melton <niklasmelton@gmail.com>"]
 license = "MIT"
+readme = "README.md"
 homepage = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 repository = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 documentation = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
-readme = "README.md"
 build = "build_extension.py"
-
-[tool.setuptools.packages.find]
-include = ["artlib*"]
-exclude = ["data*", "figures*", "templates*"]
 
 [tool.poetry.dependencies]
 python = "^3.9"
@@ -50,7 +30,13 @@ myst-parser = "^1.0"
 cffconvert = "^2.0.0"
 sphinxcontrib-bibtex = "^2.0.0"
 
-# Specify build system for C++ extensions
+[tool.poetry.urls]
+homepage = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
+
 [build-system]
-requires = ["poetry-core>=1.0.0", "setuptools", "pybind11"]
+requires = [
+    "poetry-core>=1.0.0",
+    "setuptools",          # needed for Extension(), build_ext, etc.
+    "pybind11"             # needed at build time for <pybind11.get_include()>
+]
 build-backend = "poetry.core.masonry.api"
