+++ setup.py
@@ -8,7 +8,7 @@ extra_compile_args = ["/std:c++17"] if sys.platform == "win32" else ["-std=c++17
 
 ext_modules = [
     Extension(
-        "artlib.cpp_optimized.BinaryFuzzyARTMAP",  # as you have now
+        "artlib.cpp_optimized.BinaryFuzzyARTMAP",
         [os.path.join("artlib", "cpp_optimized", "BinaryFuzzyARTMAP.cpp")],
         include_dirs=[pybind11.get_include()],
         language="c++",
