+++ pyproject.toml
@@ -4,10 +4,29 @@ version = "0.1.3"
 description = "A Python library for Adaptive Resonance Theory (ART) algorithms."
 authors = ["Niklas M. Melton <niklasmelton@gmail.com>"]
 license = "MIT"
-readme = "README.md"
 homepage = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 repository = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 documentation = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
+readme = "README.md"
+keywords = ["adaptive resonance theory", "ART", "machine learning", "neural networks", "clustering"]
+packages = [
+  { include = "artlib" },
+  { include = "artlib.cpp_optimized" }
+]
+
+classifiers = [
+  "Development Status :: 3 - Alpha",
+  "Intended Audience :: Developers",
+  "Intended Audience :: Science/Research",
+  "License :: OSI Approved :: MIT License",
+  "Operating System :: OS Independent",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Topic :: Scientific/Engineering :: Artificial Intelligence",
+]
 
 [tool.poetry.dependencies]
 python = "^3.9"
@@ -23,14 +42,11 @@ pre-commit = "^4.0.0"
 
 [tool.poetry.group.docs.dependencies]
 sphinx = "^6.2.1"
-sphinx-rtd-theme = "^1.0.0"
+sphinx-rtd-theme = "^1.0.0"  # If you're using the Read the Docs theme
 sphinx-autoapi = ">=3.0.0"
 myst-parser = "^1.0"
 cffconvert = "^2.0.0"
 sphinxcontrib-bibtex = "^2.0.0"
-
-[tool.poetry.urls]
-homepage = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 
 [build-system]
 requires = [
