+++ pyproject.toml
@@ -27,6 +27,11 @@ classifiers = [
   "Topic :: Scientific/Engineering :: Artificial Intelligence",
 ]
 
+[tool.poetry.include]
+# Force-include the .cpp and init file even if Poetry doesn't see them as standard .py
+"artlib/cpp_optimized/*.cpp" = "artlib/cpp_optimized"
+"artlib/cpp_optimized/__init__.py" = "artlib/cpp_optimized"
+
 [tool.poetry.dependencies]
 python = "^3.9"
 numpy = ">=1.19.4"
