+++ .github/workflows/pytest.yml
@@ -25,6 +25,11 @@ jobs:
         pip install poetry
         poetry install --with dev  # Install dependencies from pyproject.toml, including dev dependencies
 
+    - name: Debug Poetry Package Contents
+      run: |
+        poetry build
+        tar -tf dist/*.tar.gz | grep cpp_optimized || echo "No C++ files found!"
+
     - name: Run tests
       run: |
         poetry run pytest unit_tests/ --maxfail=5 --disable-warnings -v --tb=short --junitxml=./test-results/pytest-report.xml  # Ensure JUnit XML report is generated
