+++ pyproject.toml
@@ -12,6 +12,10 @@ keywords = ["adaptive resonance theory", "ART", "machine learning", "neural netw
 packages = [
   { include = "artlib" },
 ]
+include = [
+    "artlib/cpp_optimized/*.cpp",
+    "artlib/cpp_optimized/__init__.py"
+]
 
 classifiers = [
   "Development Status :: 3 - Alpha",
@@ -27,10 +31,6 @@ classifiers = [
   "Topic :: Scientific/Engineering :: Artificial Intelligence",
 ]
 
-[tool.poetry.include]
-# Force-include the .cpp and init file even if Poetry doesn't see them as standard .py
-"artlib/cpp_optimized/*.cpp" = "artlib/cpp_optimized"
-"artlib/cpp_optimized/__init__.py" = "artlib/cpp_optimized"
 
 [tool.poetry.dependencies]
 python = "^3.9"
