+++ .github/workflows/pytest.yml
@@ -23,7 +23,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install poetry
-        poetry install --with dev  # Install dependencies from pyproject.toml, including dev dependencies
+        poetry install --with dev --no-root
 
     - name: Build wheel
       run: |
