+++ pyproject.toml
@@ -9,10 +9,6 @@ repository = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 documentation = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 readme = "README.md"
 keywords = ["adaptive resonance theory", "ART", "machine learning", "neural networks", "clustering"]
-include = [
-    "artlib/cpp_optimized/*.cpp",
-    "artlib/cpp_optimized/__init__.py"
-]
 
 classifiers = [
   "Development Status :: 3 - Alpha",
