+++ pyproject.toml
@@ -24,7 +24,6 @@ classifiers = [
   "Topic :: Scientific/Engineering :: Artificial Intelligence",
 ]
 
-
 [tool.poetry.dependencies]
 python = "^3.9"
 numpy = ">=1.19.4"
@@ -47,8 +46,9 @@ sphinxcontrib-bibtex = "^2.0.0"
 
 [build-system]
 requires = [
-  "setuptools>=61.0",   # needed for PEP 517
+  "poetry-core>=1.0.0",
+  "setuptools>=61.0",
   "wheel",
-  "pybind11"
+  "pybind11>=2.6"
 ]
 build-backend = "setuptools.build_meta"
