### 2025-02-27 4a29588
make free function for BinaryFuzzyARTMAP fit and allow incremental learning
