+++ artlib/cpp_optimized/BinaryFuzzyARTMAP.cpp
@@ -6,6 +6,9 @@
 #include <algorithm>
 #include <cassert>
 #include <cmath>
+#include <functional>
+#include <stdexcept>
+#include <cstring>
 
 namespace py = pybind11;
 
@@ -15,16 +18,81 @@ public:
         std::vector<int> weight;
     };
 
-    BinaryFuzzyARTMAP(double rho, double alpha, std::string MT, double epsilon)
+    BinaryFuzzyARTMAP(double rho,
+                      double alpha,
+                      std::string MT,
+                      double epsilon,
+                      py::object weights = py::none(),
+                      py::object cluster_labels = py::none())
         : base_rho_(rho), alpha_(alpha), MT_(MT), epsilon_(epsilon)
     {
         // We do NOT know dim_original_ yet, so we simply set to 0 here.
         // We'll infer it in fit() from the data.
         dim_original_ = 0;
         rho_w1_       = 0;
+        bool have_weights         = !weights.is_none();
+        bool have_cluster_labels  = !cluster_labels.is_none();
+        // Must provide both or neither:
+        if (have_weights != have_cluster_labels) {
+            throw std::invalid_argument(
+                "You must provide BOTH 'weights' and 'cluster_labels' OR neither."
+            );
+        }
+        // If the user passed existing data, initialize from them.
+        if (have_weights) {
+            // We expect 'weights' to be a list of 1D arrays
+            // exactly matching how we produce `weight_arrays` in fit().
+            py::list w_list = weights.cast<py::list>();
+
+            // We also expect 'cluster_labels' to be a 1D int array
+            py::array_t<int> c_array = cluster_labels.cast<py::array_t<int>>();
+            py::buffer_info c_info   = c_array.request();
+
+            if (c_info.ndim != 1) {
+                throw std::runtime_error("cluster_labels must be a 1D array.");
+            }
+
+            py::ssize_t n_clusters = w_list.size();
+            if (c_info.shape[0] != n_clusters) {
+                throw std::runtime_error(
+                    "Inconsistent sizes: weights has " + std::to_string(n_clusters)
+                    + " clusters, but cluster_labels has length "
+                    + std::to_string(c_info.shape[0]) + "."
+                );
+            }
+
+            // Convert cluster_labels to a simple pointer
+            const int* c_ptr = static_cast<const int*>(c_info.ptr);
+
+            // Initialize clusters_ and cluster_map_
+            clusters_.resize(n_clusters);
+
+            for (py::ssize_t i = 0; i < n_clusters; ++i) {
+                // Each entry in w_list should be a 1D array.
+                py::array_t<int> w_array = w_list[i].cast<py::array_t<int>>();
+                py::buffer_info w_info   = w_array.request();
+
+                if (w_info.ndim != 1) {
+                    throw std::runtime_error("Each weight array must be 1D.");
+                }
+
+                // Copy data into clusters_[i].weight
+                py::ssize_t weight_size = w_info.shape[0];
+                clusters_[i].weight.resize(static_cast<size_t>(weight_size));
+                std::memcpy(
+                    clusters_[i].weight.data(),
+                    w_info.ptr,
+                    static_cast<size_t>(weight_size) * sizeof(int)
+                );
+
+                // cluster_map_[i] = label
+                cluster_map_[static_cast<int>(i)] = c_ptr[i];
+            }
+        }
     }
 
[…]
