### 2025-02-27 766c00f
make free function for BinaryFuzzyARTMAP fit and allow incremental learning
