### 2025-02-27 0986a38
Debug FitBinaryFuzzyARTMAP
