### 2025-02-27 3fb3703
Debug FitBinaryFuzzyARTMAP
