### 2025-02-28 77e7f68
Add PredictBinaryFuzzyARTMAP
