+++ artlib/cpp_optimized/BinaryFuzzyARTMAP.py
@@ -1,0 +1,262 @@
+"""Binary Fuzzy ART :cite:`carpenter1991fuzzy`."""
+import numpy as np
+from typing import Literal
+from artlib.cpp_optimized.cppBinaryFuzzyARTMAP import (
+    FitBinaryFuzzyARTMAP,
+    PredictBinaryFuzzyARTMAP,
+)
+from artlib.supervised.SimpleARTMAP import SimpleARTMAP
+from artlib.elementary.BinaryFuzzyART import BinaryFuzzyART
+from sklearn.utils.multiclass import unique_labels
+from sklearn.utils.validation import check_is_fitted
+
+
+class BinaryFuzzyARTMAP(SimpleARTMAP):
+    """BinaryFuzzyARTMAP for Classification. optimized with C++
+
+    This module implements BinaryFuzzyARTMAP
+
+    BinaryFuzzyARTMAP is a non-modular classification model which has been highly
+    optimized for run-time performance. Fit and predict functions are implemented in
+    c++ for efficient execution. This class acts as a wrapper for the underlying c++
+    functions and to provide compatibility with the artlib style and usage.
+    Functionally, BinaryFuzzyARTMAP behaves as a special case of
+    :class:`~artlib.supervised.SimpleARTMAP.SimpleARTMAP` instantiated with
+    :class:`~artlib.elementary.BinaryFuzzyART.BinaryFuzzyART`.
+
+    """
+
+    def __init__(self, rho: float, alpha: float):
+        """Initialize the Binary Fuzzy ARTMAP model.
+
+        Parameters
+        ----------
+        rho : float
+            Vigilance parameter.
+        alpha : float
+            Choice parameter.
+
+        """
+        module_a = BinaryFuzzyART(rho=rho, alpha=alpha)
+        super(BinaryFuzzyARTMAP, self).__init__(module_a)
+
+    def _synchronize_cpp_results(
+        self,
+        labels_a_out: np.ndarray,
+        weights_arrays: np.ndarray,
+        cluster_labels_out: np.ndarray,
+        incremental: bool = False,
+    ):
+        """Synchronize the python class with the output of the c++ code.
+
+        Parameters
+        ----------
+        labels_a_out : np.ndarray
+            A 1D numpy array containing the a-side labels from fitting
+        weights_arrays : np.ndarray
+            A 2D numpy array where rows are the Binary Fuzzy ART weights
+        cluster_labels_out : np.ndarray
+            A 1D numpy array describing the map from a-side to b-side cluster labels
+        incremental: bool, default=False
+            This flag is set to true when synchronizing after a partial_fit
+
+        """
+        if not incremental:
+            self.map: dict[int, int] = dict()
+            self.module_a.labels_ = np.array(())
+            self.module_a.weight_sample_counter_ = []
+
+        # concatenate new module_a labels to the existing labels
+        self.module_a.labels_ = np.stack(
+            [self.module_a.labels_, labels_a_out.flatten()]
+        )
+        # update module_a sample counter
+        new_sample_counts = list(map(int, np.bincount(labels_a_out)))
+
+        if len(self.module_a.weight_sample_counter_) < len(new_sample_counts):
+            self.module_a.weight_sample_counter_.extend(
+                [0]
+                * (len(new_sample_counts) - len(self.module_a.weight_sample_counter_))
+            )
+
+        for i in range(len(new_sample_counts)):
+            self.module_a.weight_sample_counter_[i] += new_sample_counts[i]
+
+        # update module_a weights
+        self.module_a.W = [w for w in weights_arrays]
+
+        # update the cluster-label map
+        for c_a, c_b in enumerate(cluster_labels_out):
+            if c_a in self.map:
+                assert self.map[c_a] == c_b, (
+                    "Incremental learning has changed "
[…]
