### 2025-03-01 5c0213f
fix cpp predict function
