### 2025-03-01 175b166
fix cpp predict function
