### 2025-03-03 53ffd85
verbose functions in cpp code
