+++ artlib/cpp_optimized/cppBinaryFuzzyARTMAP.cpp
@@ -233,6 +233,212 @@ public:
         return std::make_tuple(pred_a_out, pred_b_out);
     }
 
+    /**
+     * @brief Compute the T-value ("category choice") for each existing cluster,
+     *        re-using the same internal logic as step_fit / predict, but without
+     *        any skipping or vigilance checks.
+     *
+     * @param sample A 1D array of int features.
+     * @return 1D array of double T-values, one per cluster (could be NaN for some).
+     */
+    py::array_t<double> compute_activations_for_sample(py::array_t<int> sample) {
+        // 1) Check that we have at least one cluster
+        if (clusters_.empty()) {
+            throw std::runtime_error(
+                "Model has no clusters. Provide valid 'weights' + 'cluster_labels' first."
+            );
+        }
+
+        // 2) Validate sample is 1D
+        py::buffer_info sample_info = sample.request();
+        if (sample_info.ndim != 1) {
+            throw std::runtime_error("Expected 'sample' to be a 1D array.");
+        }
+
+        // Convert input sample into a std::vector<int>
+        auto sample_ptr = static_cast<int*>(sample_info.ptr);
+        std::vector<int> sample_vec(sample_ptr, sample_ptr + sample_info.shape[0]);
+
+        // 3) For each cluster, we call the exact same 'category_choice(...)'
+        //    that the model uses in 'step_fit(...)'.
+        //    NOTE: This might return NaN for some clusters, if vigilance fails for "MT_" logic.
+        //    But that is the same logic the class uses, so we won't second-guess it here.
+        std::vector<double> activations(clusters_.size());
+        for (size_t c = 0; c < clusters_.size(); ++c) {
+            int w1 = 0;  // dummy reference param
+            double T = category_choice(sample_vec, clusters_[c].weight, w1, MT_);
+            activations[c] = T; // possibly NaN if fails vigilance
+        }
+
+        // 4) Convert results into a py::array_t<double> to return
+        py::array_t<double> out_array(activations.size());
+        std::memcpy(out_array.mutable_data(),
+                    activations.data(),
+                    activations.size() * sizeof(double));
+
+        return out_array;
+    }
+
+    /**
+     * @brief Perform a single-sample "step_fit" but verbosely log each cluster
+     *        check in order. Returns (final_cluster_id, log_list).
+     *
+     * @param sample A 1D numpy array of ints (the input pattern).
+     * @param c_b    The label for this sample.
+     *
+     * @return (final_cluster_id, log_list)
+     *         final_cluster_id = the cluster we ended up assigning or creating
+     *         log_list         = python list of dicts describing each step
+     */
+    std::tuple<int, py::list>
+    verbose_step_fit_single_sample(py::array_t<int> sample, int c_b)
+    {
+        py::list log_list;
+
+        // ----------------------------------------------------
+        // 1) Convert the input sample to std::vector<int>
+        // ----------------------------------------------------
+        py::buffer_info buf_info = sample.request();
+        if (buf_info.ndim != 1) {
+            throw std::runtime_error("`sample` must be a 1D array.");
+        }
+        auto s_ptr = static_cast<int*>(buf_info.ptr);
+        std::vector<int> sample_vec(s_ptr, s_ptr + buf_info.shape[0]);
+
+        // If dim_original_ is 0 and we do have clusters, try to infer it
+        if (!clusters_.empty() && dim_original_ == 0) {
+            const auto &w0 = clusters_[0].weight;
+            dim_original_ = static_cast<int>(w0.size() / 2);
+            rho_w1_       = static_cast<int>(base_rho_ * dim_original_);
+        }
+
+        // ----------------------------------------------------
+        // 2) If no clusters exist => create the first cluster
+        // ----------------------------------------------------
+        if (clusters_.empty()) {
[…]
