### 2025-03-03 5287d52
break search when rho > 1.0
