+++ artlib/cpp_optimized/cppBinaryFuzzyARTMAP.cpp
@@ -234,52 +234,6 @@ public:
     }
 
     /**
-     * @brief Compute the T-value ("category choice") for each existing cluster,
-     *        re-using the same internal logic as step_fit / predict, but without
-     *        any skipping or vigilance checks.
-     *
-     * @param sample A 1D array of int features.
-     * @return 1D array of double T-values, one per cluster (could be NaN for some).
-     */
-    py::array_t<double> compute_activations_for_sample(py::array_t<int> sample) {
-        // 1) Check that we have at least one cluster
-        if (clusters_.empty()) {
-            throw std::runtime_error(
-                "Model has no clusters. Provide valid 'weights' + 'cluster_labels' first."
-            );
-        }
-
-        // 2) Validate sample is 1D
-        py::buffer_info sample_info = sample.request();
-        if (sample_info.ndim != 1) {
-            throw std::runtime_error("Expected 'sample' to be a 1D array.");
-        }
-
-        // Convert input sample into a std::vector<int>
-        auto sample_ptr = static_cast<int*>(sample_info.ptr);
-        std::vector<int> sample_vec(sample_ptr, sample_ptr + sample_info.shape[0]);
-
-        // 3) For each cluster, we call the exact same 'category_choice(...)'
-        //    that the model uses in 'step_fit(...)'.
-        //    NOTE: This might return NaN for some clusters, if vigilance fails for "MT_" logic.
-        //    But that is the same logic the class uses, so we won't second-guess it here.
-        std::vector<double> activations(clusters_.size());
-        for (size_t c = 0; c < clusters_.size(); ++c) {
-            int w1 = 0;  // dummy reference param
-            double T = category_choice(sample_vec, clusters_[c].weight, w1, MT_);
-            activations[c] = T; // possibly NaN if fails vigilance
-        }
-
-        // 4) Convert results into a py::array_t<double> to return
-        py::array_t<double> out_array(activations.size());
-        std::memcpy(out_array.mutable_data(),
-                    activations.data(),
-                    activations.size() * sizeof(double));
-
-        return out_array;
-    }
-
-    /**
      * @brief Perform a single-sample "step_fit" but verbosely log each cluster
      *        check in order. Returns (final_cluster_id, log_list).
      *
@@ -305,100 +259,98 @@ public:
         auto s_ptr = static_cast<int*>(buf_info.ptr);
         std::vector<int> sample_vec(s_ptr, s_ptr + buf_info.shape[0]);
 
-        // If dim_original_ is 0 and we do have clusters, try to infer it
+        // If dim_original_ is 0 but we have clusters, try to infer
         if (!clusters_.empty() && dim_original_ == 0) {
             const auto &w0 = clusters_[0].weight;
             dim_original_ = static_cast<int>(w0.size() / 2);
             rho_w1_       = static_cast<int>(base_rho_ * dim_original_);
         }
 
-        // ----------------------------------------------------
-        // 2) If no clusters exist => create the first cluster
-        // ----------------------------------------------------
+        // 2) If no clusters => create first
         if (clusters_.empty()) {
-            // Possibly reset rho first, in case we rely on it
             double old_rho = rho_;
             reset_rho();
 
             py::dict entry;
-            entry["stage"]        = "no_clusters_yet";
-            entry["final_action"] = "created_cluster_0";
-            entry["rho_before"]   = old_rho;
-            entry["rho_after"]    = rho_;
-
+            entry["stage"]          = "no_clusters_yet";
+            entry["final_action"]   = "created_cluster_0";
+            entry["rho_before"]     = old_rho;
+            entry["rho_after"]      = rho_;
             log_list.append(entry);
 
             clusters_.push_back({ sample_vec });
             cluster_map_[0] = c_b;
-
             return std::make_tuple(0, log_list);
         }
 
[…]
