+++ artlib/cpp_optimized/cppBinaryFuzzyARTMAP.cpp
@@ -233,168 +233,6 @@ public:
         return std::make_tuple(pred_a_out, pred_b_out);
     }
 
-    /**
-     * @brief Perform a single-sample "step_fit" but verbosely log each cluster
-     *        check in order. Returns (final_cluster_id, log_list).
-     *
-     * @param sample A 1D numpy array of ints (the input pattern).
-     * @param c_b    The label for this sample.
-     *
-     * @return (final_cluster_id, log_list)
-     *         final_cluster_id = the cluster we ended up assigning or creating
-     *         log_list         = python list of dicts describing each step
-     */
-    std::tuple<int, py::list>
-    verbose_step_fit_single_sample(py::array_t<int> sample, int c_b)
-    {
-        py::list log_list;
-
-        // ----------------------------------------------------
-        // 1) Convert the input sample to std::vector<int>
-        // ----------------------------------------------------
-        py::buffer_info buf_info = sample.request();
-        if (buf_info.ndim != 1) {
-            throw std::runtime_error("`sample` must be a 1D array.");
-        }
-        auto s_ptr = static_cast<int*>(buf_info.ptr);
-        std::vector<int> sample_vec(s_ptr, s_ptr + buf_info.shape[0]);
-
-        // If dim_original_ is 0 but we have clusters, try to infer
-        if (!clusters_.empty() && dim_original_ == 0) {
-            const auto &w0 = clusters_[0].weight;
-            dim_original_ = static_cast<int>(w0.size() / 2);
-            rho_w1_       = static_cast<int>(base_rho_ * dim_original_);
-        }
-
-        // 2) If no clusters => create first
-        if (clusters_.empty()) {
-            double old_rho = rho_;
-            reset_rho();
-
-            py::dict entry;
-            entry["stage"]          = "no_clusters_yet";
-            entry["final_action"]   = "created_cluster_0";
-            entry["rho_before"]     = old_rho;
-            entry["rho_after"]      = rho_;
-            log_list.append(entry);
-
-            clusters_.push_back({ sample_vec });
-            cluster_map_[0] = c_b;
-            return std::make_tuple(0, log_list);
-        }
-
-        // 3) reset rho
-        double old_rho = rho_;
-        reset_rho(); // sets rho_ = base_rho_
-
-        // 4) Gather T-values
-        size_t n_clusters = clusters_.size();
-        std::vector<double> T_values(n_clusters, std::nan(""));
-        std::vector<int>    w1_values(n_clusters, 0);
-
-        for (size_t i = 0; i < n_clusters; ++i) {
-            int w1;
-            double T = category_choice(sample_vec, clusters_[i].weight, w1, MT_);
-            if (!std::isnan(T)) {
-                T_values[i]  = T;
-                w1_values[i] = w1;
-            }
-        }
-
-        auto match_op = _match_tracking_operator(MT_);
-        int final_cluster_id = -1;
-
-        // 5) Repeatedly pick argmax until we find or run out
-        while (true) {
-            // find argmax
-            int best_idx = -1;
-            double best_val = -1.0;
-            for (size_t i = 0; i < n_clusters; ++i) {
-                double tv = T_values[i];
-                if (!std::isnan(tv) && tv > best_val) {
-                    best_val = tv;
-                    best_idx = static_cast<int>(i);
-                }
-            }
-
-            // If best_idx == -1 => all T are NaN => break => new cluster
-            if (best_idx < 0) {
-                break;
-            }
-
-            // build a log dict for this attempt
-            py::dict entry;
-            entry["cluster_index"]   = best_idx;
-            entry["activation"]      = T_values[best_idx];
-            entry["rho_before"]      = rho_;
-            double w1 = static_cast<double>(w1_values[best_idx]);
-            double vig_value = w1 / dim_original_;
[…]
