### 2025-03-05 49327b9
fix predict method
