+++ references.bib
@@ -53,6 +53,7 @@
   pages={493--504},
   year={1991},
   publisher={Elsevier}
+  doi={10.1016/0893-6080(91)90045-7}
 }
 @article{vigdor2007bayesian,
   title={The bayesian artmap},
