+++ references.bib
@@ -52,7 +52,7 @@
   number={4},
   pages={493--504},
   year={1991},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/0893-6080(91)90045-7}
 }
 @article{vigdor2007bayesian,
