### 2025-03-21 #7764 @chenxinye
First, I see the action test keep failing: https://github.com/NiklasMelton/AdaptiveResonanceLib/actions/runs/13958448999/job/39075007217

Do you add CI test?
