### 2025-03-21 #7764 @chenxinye
Second, do you mind put your documentation (https://adaptiveresonancelib.readthedocs.io/en/latest/index.html) in the README.md file, since it is more convenient for users to check
