### 2025-03-24 #7764 @NiklasMelton
@chenxinye I have fixed the action from failing. I had done a rushed (forced) update to the README the day prior to your comment and did not see the action fail before hand. So thank you for catching that.

I'm not sure what you mean by your second comment. I already have a README.md located in the root directory
https://github.com/NiklasMelton/AdaptiveResonanceLib/blob/develop/README.md
The README.md is meant to give usage instruction and a high-level overview. The documentation in the readthedocs is meant to give more specific details for each model. The homepage of the readthedocs actually pulls the README.md as source.
