+++ .github/workflows/pre_commit_checks.yml
@@ -38,6 +38,12 @@ jobs:
         env:
           GITHUB_BASE_REF: ${{ github.event.pull_request.base.ref }}
 
+      - name: Print base branch
+        run: |
+          echo "Base branch is: ${GITHUB_BASE_REF}"
+        env:
+          GITHUB_BASE_REF: ${{ github.event.pull_request.base.ref }}
+
       - name: Run pre-commit on changed files
         run: |
           poetry run pre-commit run --from-ref origin/${GITHUB_BASE_REF} --to-ref HEAD
