### 2025-04-02 #7764 @mstimberg
:wave: @TahiriNadia, did you find time to look at the software/paper yet? Could you please generate your checklist (`@editorialbot generate my checklist` at the top of a new comment) so that we have a rough idea of where the review's at? Thanks!
