### 2025-04-23 aa9626a
SphericalAlphaART
