+++ artlib/experimental/SphericalAlphaART.py
@@ -1,0 +1,222 @@
+import numpy as np
+from matplotlib.axes import Axes
+from typing import Optional, Iterable, List, Tuple, Dict, Callable
+from warnings import warn
+from pyalphashape import (
+    SphericalAlphaShape,
+    plot_spherical_alpha_shape,
+    latlon_to_unit_vectors
+)
+from artlib.common.utils import IndexableOrKeyable
+from artlib.experimental.AlphaART import AlphaART
+import operator
+from copy import deepcopy
+
+class AlphaART(AlphaART):
+    """
+    Alpha-Shape ART for Clustering
+    """
+
+    def __init__(self, rho: float, alpha: float):
+        """
+        Initializes the SphericalAlphaART object.
+
+        Parameters
+        ----------
+        rho : float
+            Vigilance parameter.
+        alpha : float
+            alpha shape parameter.
+        """
+        params = {"rho": rho, "alpha": alpha}
+        super().__init__(params)
+
+    @staticmethod
+    def validate_params(params: dict):
+        """
+        Validates clustering parameters.
+
+        Parameters
+        ----------
+        params : dict
+            Dictionary containing parameters for the algorithm.
+
+        """
+        assert "rho" in params
+        assert params["rho"] >= 0.0
+        assert isinstance(params["rho"], float)
+
+        assert "alpha" in params
+        assert params["alpha"] >= 0.0
+        assert isinstance(params["alpha"], float)
+
+
+    def category_choice(
+        self, i: np.ndarray, w: SphericalAlphaShape, params: dict
+    ) -> tuple[float, Optional[dict]]:
+        """
+        Get the activation of the cluster.
+
+        Parameters
+        ----------
+        i : np.ndarray
+            Data sample.
+        w : AlphaShape
+            Cluster weight or information.
+        params : dict
+            Dictionary containing parameters for the algorithm.
+
+        Returns
+        -------
+        float
+            Cluster activation.
+        dict, optional
+            Cache used for later processing.
+
+        """
+        activation = 1.0-(w.distance_to_surface(i)/np.pi)
+
+        cache = {"internal": activation >= 1.0}
+
+        return activation, cache
+
+    def match_criterion(
+        self,
+        i: np.ndarray,
+        w: SphericalAlphaShape,
+        params: dict,
+        cache: Optional[dict] = None,
+    ) -> Tuple[float, Optional[Dict]]:
+        """
+        Get the match criterion of the cluster.
+
+        Parameters
+        ----------
+        i : np.ndarray
+            Data sample.
+        w : AlphaShape
+            Cluster weight or information.
+        params : dict
+            Dictionary containing parameters for the algorithm.
+        cache : dict, optional
+            Cache containing values cached from previous calculations.
+
+        Returns
+        -------
+        float
+            Cluster match criterion.
+        dict
+            Cache used for later processing.
+
+        """
+        assert cache is not None and "internal" in cache
+        if cache["internal"]:
+            M = 1.0
+        else:
+            pt = latlon_to_unit_vectors(i)
+            dot_products = np.clip(w.perimeter_points @ pt, -1.0, 1.0)
+            angles = np.arccos(dot_products)
+            dist = np.max(angles)
+            M = 1.0 - (dist/np.pi)
+
+        return M, cache
+
+    def new_weight(self, i: np.ndarray, params: dict) -> SphericalAlphaShape:
+        """
+        Generate a new cluster weight.
+
+        Parameters
+        ----------
+        i : np.ndarray
+            Data sample.
+        params : dict
+            Dictionary containing parameters for the algorithm.
+
+        Returns
+        -------
+        AlphaShape
+            New cluster weight.
[…]
