### 2025-04-23 0c5e725
SphericalAlphaART
