### 2025-04-23 68e6da0
SphericalAlphaART doc strings
