### 2025-04-23 fa323ae
SphericalAlphaART doc strings
