### 2025-04-23 f21c1a0
better check on resonance
