### 2025-05-21 #7764 @TahiriNadia
> 👋 [@TahiriNadia](https://github.com/TahiriNadia), did you find time to look at the software/paper yet? Could you please generate your checklist (`@editorialbot generate my checklist` at the top of a new comment) so that we have a rough idea of where the review's at? Thanks!

Thank you for this note and sorry for my delay.
