### 2025-05-21 842c8c0
spherical art
