### 2025-05-21 cbb6915
Spherical ART
