### 2025-05-21 #7764 @NiklasMelton
> **On substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
>
> I see the next criterion. [@NiklasMelton](https://github.com/NiklasMelton), would it be possible to provide us with examples of previous articles where this software has been cited? `Whether the software is sufficiently useful that it is likely to be cited by your peer group.` (source: [On substantial scholarly] efforthttps://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

This software has been cited in the following articles:

"An Extensive Analysis of Match-Tracking Methods for ARTMAP" https://ieeexplore.ieee.org/document/10969482
"Deep ARTMAP: Generalized Hierarchical Learning with Adaptive Resonance Theory" https://arxiv.org/abs/2503.07641

The software is still relatively new so additional articles will come in time. That said, the algorithms therein have a very extensive publication history with very few open source solutions:
"A survey of adaptive resonance theory neural network models for engineering applications" https://dl.acm.org/doi/10.1016/j.neunet.2019.09.012
