### 2025-05-21 #7764 @NiklasMelton
> Generally I am very happy with the software, just to confirm: do you compare your methods in your package with other libraries? I would like to know if they can generate unified output, and what cause the difference if any.

I have not taken the time to compare the software with others as the existing solutions are so limited and narrow in scope. At best, I would be able to compare results on a couple of special case applications. Further, none of the other implementations have any sort of notoriety or bona fides that give me confidence in them. However, I have had this work reviewed by the progenitors of the algorithms, Dr. Stephen Grossberg and Dr. Gail Carpenter, as well as several other experts in these algorithms including the authors of the survey paper linked in the above comment; All of whom are happy with the work.

That said, if the comparison is a sticking point for you, I will find those special cases and attempt a comparison.
