### 2025-05-21 #7764 @NiklasMelton
> I think would be nice if you can add more details on what is Adaptive Resonance Theory, that will help the beginners a lot. [@NiklasMelton](https://github.com/NiklasMelton)

Will do!
