### 2025-05-27 a42d5a4
Fix Spherical Alpha ART
