### 2025-05-29 9b49fd2
add proper match tracking to fusion ART
