+++ artlib/cpp_optimized/FuzzyARTMAP.py
@@ -1,0 +1,189 @@
+"""Fuzzy ARTMAP – C++‑accelerated wrapper."""
+from __future__ import annotations
+
+import numpy as np
+from typing import Literal
+
+from artlib.cpp_optimized.cppFuzzyARTMAP import (  # ← new backend
+    FitFuzzyARTMAP,
+    PredictFuzzyARTMAP,
+)
+from artlib.supervised.SimpleARTMAP import SimpleARTMAP
+from artlib.elementary.FuzzyART import FuzzyART  # ← analogue of BinaryFuzzyART
+from sklearn.utils.multiclass import unique_labels
+from sklearn.utils.validation import check_is_fitted
+
+
+class FuzzyARTMAP(SimpleARTMAP):
+    """C++‑optimised Fuzzy ARTMAP.
+
+    A thin wrapper around the `cppFuzzyARTMAP` free functions that keeps
+    the Python‑side state (`module_a`, label maps, sample counters) in sync
+    with the C++ core.
+
+    """
+
+    # ---------------------------------------------------------------------
+    # construction
+    # ---------------------------------------------------------------------
+    def __init__(self, rho: float, alpha: float, beta: float):
+        """
+        Parameters
+        ----------
+        rho   : float  – vigilance (0 ≤ ρ ≤ 1)
+        alpha : float  – choice (α ≥ 0)
+        beta  : float  – learning‑rate (0 < β ≤ 1)
+        """
+        module_a = FuzzyART(rho=rho, alpha=alpha, beta=beta)
+        super().__init__(module_a)
+
+    # ---------------------------------------------------------------------
+    # helper – mirror C++ results back to Python objects
+    # ---------------------------------------------------------------------
+    def _sync_cpp(
+        self,
+        labels_a_out: np.ndarray,
+        weights_arrays: list[np.ndarray],
+        cluster_labels_out: np.ndarray,
+        *,
+        incremental: bool = False,
+    ):
+        if not incremental:
+            self.map: dict[int, int] = {}
+            self.module_a.labels_ = np.array((), dtype=int)
+            self.module_a.weight_sample_counter_ = []
+
+        # labels
+        self.module_a.labels_ = np.concatenate(
+            [self.module_a.labels_, labels_a_out.astype(int)]
+        )
+
+        # sample counters
+        new_counts = np.bincount(labels_a_out, minlength=len(weights_arrays))
+        if len(self.module_a.weight_sample_counter_) < len(new_counts):
+            self.module_a.weight_sample_counter_.extend(
+                [0] * (len(new_counts) - len(self.module_a.weight_sample_counter_))
+            )
+        for k, c in enumerate(new_counts):
+            self.module_a.weight_sample_counter_[k] += int(c)
+
+        # weights (float64 arrays)
+        self.module_a.W = [w for w in weights_arrays]
+
+        # A→B mapping
+        for c_a, c_b in enumerate(cluster_labels_out):
+            if c_a in self.map:
+                assert self.map[c_a] == c_b, "Incremental fit changed cluster map."
+            else:
+                self.map[c_a] = int(c_b)
+
+    # ---------------------------------------------------------------------
+    # batch fit
+    # ---------------------------------------------------------------------
+    def fit(
+        self,
+        X: np.ndarray,
+        y: np.ndarray,
+        *,
+        max_iter: int = 1,
+        match_tracking: Literal["MT+", "MT-", "MT0", "MT1", "MT~"] = "MT+",
+        epsilon: float = 1e-10,
+        verbose: bool = False,
+        leave_progress_bar: bool = True,
+    ):
+        SimpleARTMAP.validate_data(self, X, y)
+        self.classes_ = unique_labels(y)
+        self.labels_ = y
+        self.module_a.W = []
[…]
