+++ artlib/cpp_optimized/cppFuzzyARTMAP.cpp
@@ -1,0 +1,381 @@
+// cppFuzzyARTMAP.cpp
+// ----------------------------------------------------------
+//  C++ accelerated Fuzzy  ARTMAP  (pybind11)
+//  ‑ supports incremental training via external weights/map
+//  ‑ real‑valued inputs in [0,1] (already normalised & complement‑coded)
+// ----------------------------------------------------------
+#include <pybind11/pybind11.h>
+#include <pybind11/numpy.h>
+#include <pybind11/stl.h>
+
+#include <vector>
+#include <unordered_map>
+#include <algorithm>
+#include <cassert>
+#include <cmath>
+#include <functional>
+#include <limits>
+#include <stdexcept>
+#include <cstring>
+
+namespace py = pybind11;
+
+// ────────────────────────────────────────────────────────────
+class cppFuzzyARTMAP {
+public:
+    struct Cluster {
+        std::vector<double> weight;          // 2d components (complement‑coded)
+    };
+
+    cppFuzzyARTMAP(double rho,
+                   double alpha,
+                   double beta,
+                   const std::string& MT,
+                   double epsilon,
+                   py::object weights         = py::none(),
+                   py::object cluster_labels  = py::none())
+        : base_rho_(rho),
+          alpha_(alpha),
+          beta_(beta),
+          MT_(MT),
+          epsilon_(epsilon),
+          dim_original_(0),
+          rho_(rho)
+    {
+        const bool have_W  = !weights.is_none();
+        const bool have_cl = !cluster_labels.is_none();
+        if (have_W != have_cl) {
+            throw std::invalid_argument(
+                "Provide BOTH 'weights' and 'cluster_labels' or neither.");
+        }
+
+        if (have_W) {
+            py::list      w_list = weights.cast<py::list>();
+            py::array_t<int> cl  = cluster_labels.cast<py::array_t<int>>();
+            auto            cl_b = cl.request();
+            if (cl_b.ndim != 1)
+                throw std::runtime_error("cluster_labels must be 1‑D");
+
+            const std::size_t n_clusters = w_list.size();
+            if (static_cast<std::size_t>(cl_b.shape[0]) != n_clusters)
+                throw std::runtime_error("weights / cluster_labels size mismatch");
+
+            clusters_.resize(n_clusters);
+            const int* cl_ptr = static_cast<const int*>(cl_b.ptr);
+
+            for (std::size_t k = 0; k < n_clusters; ++k) {
+                py::array_t<double> w_arr = w_list[k].cast<py::array_t<double>>();
+                auto                w_b   = w_arr.request();
+                if (w_b.ndim != 1)
+                    throw std::runtime_error("each weight must be 1‑D");
+                clusters_[k].weight.resize(static_cast<std::size_t>(w_b.shape[0]));
+                std::memcpy(clusters_[k].weight.data(),
+                            w_b.ptr,
+                            sizeof(double) * w_b.shape[0]);
+                cluster_map_[static_cast<int>(k)] = cl_ptr[k];
+
+                if (dim_original_ == 0) dim_original_ = static_cast<int>(w_b.shape[0] / 2);
+            }
+        }
+    }
+
+    // ────────────────────────────────────────────────────────
+    // FIT
+    // returns (labels_a, weights_vec<ndarray>, cluster_labels)
+    // ────────────────────────────────────────────────────────
+    std::tuple<py::array_t<int>,
+               std::vector<py::array_t<double>>,
+               py::array_t<int>>
+    fit(py::array_t<double> X, py::array_t<int> y)
+    {
+        auto xb = X.request();
+        auto yb = y.request();
+
+        assert(xb.ndim == 2 && yb.ndim == 1);
[…]
