+++ artlib/cpp_optimized/FuzzyARTMAP.py
@@ -1,9 +1,6 @@
-"""Fuzzy ARTMAP – C++‑accelerated wrapper."""
-from __future__ import annotations
-
+"""Fuzzy ARTMAP :cite:`carpenter1991fuzzy`."""
 import numpy as np
 from typing import Literal
-
 from artlib.cpp_optimized.cppFuzzyARTMAP import (  # ← new backend
     FitFuzzyARTMAP,
     PredictFuzzyARTMAP,
@@ -15,39 +12,57 @@ from sklearn.utils.validation import check_is_fitted
 
 
 class FuzzyARTMAP(SimpleARTMAP):
-    """C++‑optimised Fuzzy ARTMAP.
+    """FuzzyARTMAP for Classification. optimized with C++
 
-    A thin wrapper around the `cppFuzzyARTMAP` free functions that keeps
-    the Python‑side state (`module_a`, label maps, sample counters) in sync
-    with the C++ core.
+    This module implements FuzzyARTMAP
+
+    FuzzyARTMAP is a non-modular classification model which has been highly
+    optimized for run-time performance. Fit and predict functions are implemented in
+    c++ for efficient execution. This class acts as a wrapper for the underlying c++
+    functions and to provide compatibility with the artlib style and usage.
+    Functionally, FuzzyARTMAP behaves as a special case of
+    :class:`~artlib.supervised.SimpleARTMAP.SimpleARTMAP` instantiated with
+    :class:`~artlib.elementary.FuzzyART.FuzzyART`.
 
     """
 
-    # ---------------------------------------------------------------------
-    # construction
-    # ---------------------------------------------------------------------
     def __init__(self, rho: float, alpha: float, beta: float):
-        """
+        """Initialize the Binary Fuzzy ARTMAP model.
+
         Parameters
         ----------
-        rho   : float  – vigilance (0 ≤ ρ ≤ 1)
-        alpha : float  – choice (α ≥ 0)
-        beta  : float  – learning‑rate (0 < β ≤ 1)
+        rho : float
+            Vigilance parameter.
+        alpha : float
+            Choice parameter.
+        beta : float
+            Learning rate.
+
         """
         module_a = FuzzyART(rho=rho, alpha=alpha, beta=beta)
         super().__init__(module_a)
 
-    # ---------------------------------------------------------------------
-    # helper – mirror C++ results back to Python objects
-    # ---------------------------------------------------------------------
-    def _sync_cpp(
+    def _synchronize_cpp_results(
         self,
         labels_a_out: np.ndarray,
         weights_arrays: list[np.ndarray],
         cluster_labels_out: np.ndarray,
-        *,
         incremental: bool = False,
     ):
+        """Synchronize the python class with the output of the c++ code.
+
+        Parameters
+        ----------
+        labels_a_out : np.ndarray
+            A 1D numpy array containing the a-side labels from fitting
+        weights_arrays : np.ndarray
+            A 2D numpy array where rows are the Binary Fuzzy ART weights
+        cluster_labels_out : np.ndarray
+            A 1D numpy array describing the map from a-side to b-side cluster labels
+        incremental: bool, default=False
+            This flag is set to true when synchronizing after a partial_fit
+
+        """
         if not incremental:
             self.map: dict[int, int] = {}
             self.module_a.labels_ = np.array((), dtype=int)
@@ -77,20 +92,41 @@ class FuzzyARTMAP(SimpleARTMAP):
             else:
                 self.map[c_a] = int(c_b)
 
-    # ---------------------------------------------------------------------
-    # batch fit
-    # ---------------------------------------------------------------------
     def fit(
         self,
         X: np.ndarray,
         y: np.ndarray,
-        *,
         max_iter: int = 1,
[…]
