+++ docs/source/what_is_art.rst
@@ -1,0 +1,5 @@
+.. include:: ../../README.md
+   :parser: myst_parser.sphinx_
+   :start-after: <!-- START what is ART -->
+   :end-before: <!-- END what is ART -->
+
