### 2025-06-10 80e6196
add Hypersphere ARTMAP and FuzzyARTMAP unit test
