### 2025-06-10 4e5a3b7
add Hypersphere ARTMAP and FuzzyARTMAP unit test
