### 2025-06-10 84d9778
Add hypersphereARTMAP unit tests
