### 2025-06-11 f94fcf8
add GaussianARTMAP cpp code and improve ARTMAP unit tests
