+++ artlib/__init__.py
@@ -55,6 +55,7 @@ from artlib.topological.DualVigilanceART import DualVigilanceART
 from artlib.cpp_optimized.BinaryFuzzyARTMAP import BinaryFuzzyARTMAP
 from artlib.cpp_optimized.FuzzyARTMAP import FuzzyARTMAP
 from artlib.cpp_optimized.HypersphereARTMAP import HypersphereARTMAP
+from artlib.cpp_optimized.GaussianARTMAP import GaussianARTMAP
 
 __all__ = [
     "BaseART",
@@ -88,4 +89,5 @@ __all__ = [
     "BinaryFuzzyARTMAP",
     "FuzzyARTMAP",
     "HypersphereARTMAP",
+    "GaussianARTMAP",
 ]
