+++ unit_tests/test_GaussianARTMAP.py
@@ -1,9 +1,9 @@
 import pytest
 import numpy as np
 from sklearn.datasets import make_blobs
-from artlib.elementary.FuzzyART import FuzzyART
+from artlib.elementary.GaussianART import GaussianART
 from artlib.supervised.SimpleARTMAP import SimpleARTMAP
-from artlib.cpp_optimized.FuzzyARTMAP import FuzzyARTMAP
+from artlib.cpp_optimized.GaussianARTMAP import GaussianARTMAP
 
 
 def test_prepare_data():
@@ -14,9 +14,13 @@ def test_prepare_data():
         random_state=0,
         shuffle=False,
     )
-    params = {"rho": 0.8, "alpha": 1e-10, "beta": 1.0}
-    A = SimpleARTMAP(FuzzyART(**params))
-    B = FuzzyARTMAP(**params)
+    params = {
+        "rho": 0.05,
+        "alpha": 1e-10,
+        "sigma_init": np.array([0.5, 0.5]),
+    }
+    A = SimpleARTMAP(HypersphereART(**params))
+    B = GaussianARTMAP(**params)
 
     X_A = A.prepare_data(data)
     X_B = B.prepare_data(data)
@@ -33,9 +37,13 @@ def test_consistency():
             shuffle=False,
         )
 
-    params = {"rho":0.8, "alpha":1e-10, "beta":1.0}
-    A = SimpleARTMAP(FuzzyART(**params))
-    B = FuzzyARTMAP(**params)
+    params = {
+        "rho": 0.05,
+        "alpha": 1e-10,
+        "sigma_init": np.array([0.5, 0.5]),
+    }
+    A = SimpleARTMAP(GaussianART(**params))
+    B = GaussianARTMAP(**params)
 
     X = A.prepare_data(data)
 
