### 2025-06-11 566610f
add GaussianARTMAP cpp code and improve ARTMAP unit tests
