### 2025-06-11 8a82820
add binary artmap unit test
