### 2025-06-11 d190b69
add binary art prepare data method
