+++ artlib/experimental/iAlphaShape.py
@@ -1,382 +1,0 @@
-import itertools
-import numpy as np
-from typing import Tuple, Set
-import heapq
-from scipy.spatial import ConvexHull, KDTree
-from scipy.spatial.distance import cdist
-from artlib.experimental.AlphaShape import AlphaShape
-
-def max_volume_seed(points: np.ndarray) -> np.ndarray:
-    """
-    Return indices of a (d+1)-simplex with near‑maximal volume using a
-    greedy farthest‑point heuristic.
-    """
-    pts = points
-    d   = pts.shape[1]
-    N   = len(pts)
-
-    # 1. start with the global centroid
-    centroid = pts.mean(axis=0, keepdims=True)
-    first    = np.argmax(np.linalg.norm(pts - centroid, axis=1))
-    seed     = [first]
-
-    # 2. farthest‑point iterations
-    while len(seed) < d + 1:
-        dist2 = np.min(cdist(pts, pts[seed])**2, axis=1)
-        seed.append(int(np.argmax(dist2)))
-
-    return np.array(seed, dtype=int)
-def presort_for_alpha(points: np.ndarray,
-                      alpha: float,
-                      seed_layer: int = 1) -> np.ndarray:
-    """
-    Return a reordered copy of `points` that is well‑suited for incremental
-    α‑shape construction.
-
-    Parameters
-    ----------
-    points      : (N,d) array
-    alpha       : α used for the shape (needed for the distance window)
-    seed_layer  : how many hull layers to peel off before incremental insert
-
-    Returns
-    -------
-    points_sorted : (N,d) array
-    """
-    pts = points.copy()
-    d   = pts.shape[1]
-    r_filter = np.inf if alpha <= 0 else 1.0 / alpha
-
-    seed_idx = []
-    remaining = np.arange(len(pts))
-
-    # -------- Stage 1 : convex‑hull peeling --------------------------
-    for _ in range(seed_layer):
-        if len(remaining) < d + 1:
-            break
-        hull = ConvexHull(pts[remaining])
-        layer = remaining[hull.vertices]
-        seed_idx.extend(layer.tolist())
-        remaining = np.setdiff1d(remaining, layer, assume_unique=True)
-
-    if len(seed_idx) < d + 1:
-        # fall back to random if hull failed (degenerate cloud)
-        extra = remaining[: (d + 1 - len(seed_idx))]
-        seed_idx.extend(extra.tolist())
-        remaining = np.setdiff1d(remaining, extra, assume_unique=True)
-
-    # -------- Stage 2 : sort by distance to hull ---------------------
-    if len(remaining):
-        hull_pts = pts[seed_idx]
-        tree = KDTree(hull_pts)
-        dists, _ = tree.query(pts[remaining])
-        order = np.argsort(dists)          # closest‑to‑hull first
-        remaining = remaining[order]
-
-    # -------- concatenate & return -----------------------------------
-    return pts[np.concatenate([seed_idx, remaining])]
-
-
-def face_is_visible(all_points: np.ndarray,
-                    face_idx: Tuple[int, ...],
-                    new_pt: np.ndarray,
-                    interior_ref: np.ndarray,
-                    eps: float = 1e-12) -> bool:
-    """
-    Return *True* iff `new_pt` lies on the opposite side of the hyper‑plane
-    that supports the boundary face `face_idx`, compared to an interior
-    reference point.
-
-    Parameters
-    ----------
-    all_points   : (N,d) array holding every vertex so far.
-    face_idx     : tuple of length (d) – indices of the (d‑1)-simplex that
-                   forms the boundary face.
-    new_pt       : (d,) array – the point being inserted.
[…]
