### 2025-06-11 c814ef7
remove testing scripts
