+++ plot_hull_art_results.py
@@ -1,966 +1,0 @@
-import matplotlib.pyplot as plt
-import matplotlib.patches as patches
-import matplotlib.lines as mlines
-import matplotlib.scale as mscale
-import matplotlib.transforms as mtransforms
-import matplotlib.ticker as ticker
-import colorspacious as cs
-from scipy import stats
-import seaborn as sns
-import pandas as pd
-import numpy as np
-from matplotlib.transforms import Affine2D
-
-import warnings
-
-warnings.filterwarnings("ignore", category=DeprecationWarning)
-
-
-class CustomSymLogScale(mscale.ScaleBase):
-    """Custom symmetric log scale that behaves logarithmically for large values but is
-    linear between -linthresh and +linthresh."""
-
-    name = "custom_symlog"
-
-    def __init__(self, axis, *, linthresh=1e-5, base=10, subs=None, **kwargs):
-        # Properly call the parent class constructor
-        super().__init__(axis)
-        self.linthresh = linthresh
-        self.base = base
-        self.subs = subs
-
-    def get_transform(self):
-        return self.CustomSymLogTransform(self.linthresh, self.base)
-
-    def set_default_locators_and_formatters(self, axis):
-        axis.set_major_locator(plt.LogLocator(self.base, subs=self.subs))
-        axis.set_minor_locator(plt.LogLocator(self.base, subs=self.subs, numticks=12))
-        axis.set_major_formatter(plt.ScalarFormatter())
-
-    class CustomSymLogTransform(mtransforms.Transform):
-        input_dims = output_dims = 1
-
-        def __init__(self, linthresh, base):
-            mtransforms.Transform.__init__(self)
-            self.linthresh = linthresh
-            self.log_base = np.log(base)
-
-        def transform_non_affine(self, values):
-            """Apply custom symmetric log transformation."""
-            sign = np.sign(values)
-            abs_values = np.abs(values)
-
-            with np.errstate(invalid="ignore"):
-                return sign * np.where(
-                    abs_values < self.linthresh,
-                    abs_values,  # Linear transformation below the threshold
-                    self.linthresh
-                    + np.log(np.maximum(abs_values / self.linthresh, 1e-50))
-                    / self.log_base,
-                )
-
-        def inverted(self):
-            return CustomSymLogScale.InvertedCustomSymLogTransform(
-                self.linthresh, self.log_base
-            )
-
-    class InvertedCustomSymLogTransform(mtransforms.Transform):
-        input_dims = output_dims = 1
-
-        def __init__(self, linthresh, log_base):
-            mtransforms.Transform.__init__(self)
-            self.linthresh = linthresh
-            self.log_base = log_base
-
-        def transform_non_affine(self, values):
-            """Apply inverse of custom symmetric log transformation."""
-            sign = np.sign(values)
-            abs_values = np.abs(values)
-
-            return sign * np.where(
-                abs_values < self.linthresh,
-                abs_values,
-                self.linthresh * np.exp((abs_values - self.linthresh) * self.log_base),
-            )
-
-        def inverted(self):
-            return CustomSymLogScale.CustomSymLogTransform(
-                self.linthresh, self.log_base
-            )
-
-
-# Register the scale with matplotlib
-mscale.register_scale(CustomSymLogScale)
-
-
-def smart_title(input_string):
-    # Apply title() to the string
-    title_string = input_string.title()
-
-    # Create a new string by keeping the original uppercase where applicable
-    result = "".join(
-        [
[…]
