### 2025-06-12 a548ba7
add cpp docs
