+++ docs/source/cpp_optimized.rst
@@ -1,0 +1,5 @@
+.. include:: ../../README.md
+   :parser: myst_parser.sphinx_
+   :start-after: <!-- START cpp_optimized -->
+   :end-before: <!-- END cpp_optimized -->
+
