### 2025-06-12 d378266
add cpp docs
