+++ README.md
@@ -132,6 +132,7 @@ AdaptiveResonanceLib includes implementations for the following ART models:
     - [Ellipsoid ART](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.html#artlib.EllipsoidART)
     - [Fuzzy ART](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.html#artlib.FuzzyART)
     - [Quadratic Neuron ART](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.html#artlib.QuadraticNeuronART)
+    - [Binary Fuzzy ART](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.html#artlib.BinaryFuzzyART)
 - #### Metric Informed
     - [CVI ART](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.html#artlib.CVIART)
     - [iCVI Fuzzy ART](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.html#artlib.iCVIFuzzyART)
@@ -153,6 +154,12 @@ AdaptiveResonanceLib includes implementations for the following ART models:
     - [TD-FALCON](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.html#artlib.TD_FALCON)
 - #### Biclustering
     - [Biclustering ARTMAP](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.biclustering.html#artlib.biclustering.BARTMAP.BARTMAP)
+- #### C++ Accelerated
+    - [Fuzzy ARTMAP](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.cpp_optimized.html#artlib.cpp_optimized.FuzzyARTMAP.FuzzyARTMAP)
+    - [Hypersphere ARTMAP](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.cpp_optimized.html#artlib.cpp_optimized.HypersphereARTMAP.HypersphereARTMAP)
+    - [Gaussian ARTMAP](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.cpp_optimized.html#artlib.cpp_optimized.GaussianARTMAP.GaussianARTMAP)
+    - [Binary Fuzzy ARTMAP](https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.cpp_optimized.html#artlib.cpp_optimized.BinaryFuzzyARTMAP.BinaryFuzzyARTMAP)
+    -
 
 <!-- END available_models -->
 
@@ -298,6 +305,7 @@ model.fit(train_Xy_prep)
 pred_y = model.predict_regression(test_Xy_prep, target_channels=[1], clip=True)
 ```
 
+
 ### Data Normalization
 
 AdaptiveResonanceLib models require feature data to be normalized between 0.0
@@ -329,6 +337,93 @@ If only the boundaries of your testing data are unknown, you can call
 training. Only use this if you understand what you are doing.
 
 <!-- END quick-start -->
+
+<!-- START cpp_optimized -->
+
+[//]: # (## C++ Optimization)
+
+[//]: # ()
+[//]: # ( While all classes use NumPy and SciPy for mathematical operations to accelerate )
+
+[//]: # ( execution times, Some classes have Numba optimized activation and )
+
+[//]: # ( vigilance functions to further accelerate them. These include:)
+
+[//]: # ()
+[//]: # (- [ART1]&#40;https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.html#artlib.ART1&#41;)
+
+[//]: # (- [Fuzzy ART]&#40;https://adaptiveresonancelib.readthedocs.io/en/latest/artlib.html#artlib.FuzzyART&#41;)
+
[…]
