### 2025-06-12 d5dade7
add images to readme and cpp_optimized rst page
