+++ docs/source/artlib.cpp_optimized.rst
@@ -1,0 +1,47 @@
+artlib.cpp_optimized package
+=====================
+
+Module contents
+---------------
+
+.. automodule:: artlib.cpp_optimized
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+artlib.cpp_optimized.BaseART module
+----------------------------
+
+.. automodule:: artlib.cpp_optimized.BinaryFuzzyARTMAP
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+artlib.cpp_optimized.FuzzyARTMAP module
+-------------------------------
+
+.. automodule:: artlib.cpp_optimized.FuzzyARTMAP
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+artlib.cpp_optimized.HypersphereARTMAP module
+------------------------
+
+.. automodule:: artlib.cpp_optimized.HypersphereARTMAP
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+artlib.cpp_optimized.GaussianARTMAP module
+--------------------------
+
+.. automodule:: artlib.cpp_optimized.GaussianARTMAP
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
