### 2025-06-12 f0a85ae
add images to readme and cpp_optimized rst page
