### 2025-06-12 9c2197f
update cpp_optimized rst
