+++ docs/source/artlib.cpp_optimized.rst
@@ -1,47 +1,50 @@
 artlib.cpp_optimized package
-=====================
-
-Module contents
----------------
+============================
 
 .. automodule:: artlib.cpp_optimized
    :members:
    :undoc-members:
    :show-inheritance:
 
-Submodules
-----------
+Sub-modules
+-----------
 
-artlib.cpp_optimized.BaseART module
+artlib.cpp_optimized.BaseART
 ----------------------------
+
+.. automodule:: artlib.cpp_optimized.BaseART
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+artlib.cpp_optimized.BinaryFuzzyARTMAP
+--------------------------------------
 
 .. automodule:: artlib.cpp_optimized.BinaryFuzzyARTMAP
    :members:
    :undoc-members:
    :show-inheritance:
 
-artlib.cpp_optimized.FuzzyARTMAP module
--------------------------------
+artlib.cpp_optimized.FuzzyARTMAP
+--------------------------------
 
 .. automodule:: artlib.cpp_optimized.FuzzyARTMAP
    :members:
    :undoc-members:
    :show-inheritance:
 
-artlib.cpp_optimized.HypersphereARTMAP module
-------------------------
+artlib.cpp_optimized.HypersphereARTMAP
+--------------------------------------
 
 .. automodule:: artlib.cpp_optimized.HypersphereARTMAP
    :members:
    :undoc-members:
    :show-inheritance:
 
-artlib.cpp_optimized.GaussianARTMAP module
---------------------------
+artlib.cpp_optimized.GaussianARTMAP
+-----------------------------------
 
 .. automodule:: artlib.cpp_optimized.GaussianARTMAP
    :members:
    :undoc-members:
    :show-inheritance:
-
-
