### 2025-06-12 b0e82a6
add mock functions and __init__ docstring
