+++ docs/source/conf.py
@@ -45,6 +45,13 @@ autoapi_ignore = ['*/experimental', '*/experimental/*']
 # autoapi_python_class_content = 'both'
 # autoclass_content = 'both'
 
+autodoc_mock_imports = [
+    "artlib.cpp_optimized.cppBinaryFuzzyARTMAP",
+    "artlib.cpp_optimized.cppFuzzyARTMAP",
+    "artlib.cpp_optimized.cppGaussianARTMAP",
+    "artlib.cpp_optimized.cppHypersphereARTMAP",
+]
+
 bibtex_bibfiles = ['artlib_citation.bib', '../../references.bib']
 
 intersphinx_mapping = {
