### 2025-06-12 401a790
add mock functions and __init__ docstring
