### 2025-06-12 bd036de
fix PyPi description
