### 2025-06-12 48fd99a
fix PyPi description
