+++ pyproject.toml
@@ -1,3 +1,10 @@
+[project]
+name        = "artlib"
+version     = "0.1.4"
+description = "A Python library for Adaptive Resonance Theory (ART) algorithms."
+readme      = { file = "README.md", "content-type" = "text/markdown" }
+requires-python = ">=3.9"
+
 [tool.poetry]
 name = "artlib"
 version = "0.1.4"
@@ -7,7 +14,7 @@ license = "MIT"
 homepage = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 repository = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
 documentation = "https://github.com/NiklasMelton/AdaptiveResonanceLib"
-readme = { file = "README.md" }
+readme = "README.md"
 keywords = ["adaptive resonance theory", "ART", "machine learning", "neural networks", "clustering"]
 
 classifiers = [
