+++ .github/workflows/test-pypi-publish.yml
@@ -1,0 +1,60 @@
+name: Test Upload Python Package
+
+on:
+  push:
+    branches: [main]          # fires for merges or direct pushes to main
+
+permissions:
+  contents: read              # needed by checkout; nothing broader
+
+jobs:
+  deploy:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.x'
+
+      # ---------- Optional build-speed caches ----------
+      - name: Restore ccache
+        uses: actions/cache@v4
+        with:
+          path: /root/.ccache
+          key:  ccache-${{ runner.os }}-${{ hashFiles('setup.cfg', 'pyproject.toml', '**/*.[ch]pp', '**/*.[ch]') }}
+          restore-keys: ccache-${{ runner.os }}-
+
+      - name: Restore pip cache
+        uses: actions/cache@v4
+        with:
+          path: ~/.cache/pip
+          key:  pip-${{ runner.os }}-${{ hashFiles('pyproject.toml') }}
+          restore-keys: pip-${{ runner.os }}-
+      # --------------------------------------------------
+
+      - name: Install build tooling
+        run: |
+          python -m pip install --upgrade pip
+          pip install build cibuildwheel
+
+      - name: Build wheels and sdist
+        env:
+          # enable compiler caching inside cibuildwheel
+          CIBW_ENVIRONMENT: CCACHE_DIR=/root/.ccache
+          CIBW_BEFORE_ALL_LINUX: |
+            yum install -y ccache
+            export CC="ccache gcc" CXX="ccache g++"
+        run: |
+          python -m cibuildwheel --output-dir dist
+          python -m build --sdist -o dist
+
+      - name: Upload to TestPyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          repository-url: https://test.pypi.org/legacy/
+          user: __token__
+          password: ${{ secrets.TEST_PYPI_API_TOKEN }}
+          skip-existing: true            # avoids failures on re-runs
