+++ examples/demo_reinforcement_learning.py
@@ -6,6 +6,9 @@ from collections import defaultdict
 import matplotlib.pyplot as plt
 from copy import deepcopy
 
+STATE_MAX = 47
+ACTION_MAX = 3
+REWARD_MAX = 150
 
 # only works with Fuzzy ART based FALCON
 def prune_clusters(cls):
@@ -182,9 +185,6 @@ def demo_cycle(cls, epochs, steps, render_mode=None):
     env = gym.make("CliffWalking-v0", render_mode=render_mode)
     # define some constants
     ACTION_SPACE = np.array([[0], [1.0], [2.0], [3.0]])
-    STATE_MAX = 47
-    ACTION_MAX = 3
-    REWARD_MAX = 150
 
     # track reward history
     reward_history = []
@@ -287,6 +287,9 @@ def train_FALCON():
     art_state = FuzzyART(rho=0.99, alpha=0.01, beta=1.0)
     art_action = FuzzyART(rho=0.99, alpha=0.01, beta=1.0)
     art_reward = FuzzyART(rho=0.0, alpha=0.01, beta=1.0)
+    art_state.set_data_bounds(np.zeros((1,)), np.ones((1,)))
+    art_action.set_data_bounds(np.zeros((1,)), ACTION_MAX*np.ones((1,)))
+    art_reward.set_data_bounds(np.zeros((1,)), np.ones((1,)))
     # instantiate FALCON
     cls = TD_FALCON(art_state, art_action, art_reward, channel_dims=[2, 2, 2])
     # record rewards for each epoch
