### 2025-06-27 a711b26
fix partial fit issue with cpp ARTMAP
