+++ artlib/common/BaseART.py
@@ -31,6 +31,7 @@ class BaseART(BaseEstimator, ClusterMixin):
         self.d_min_ = None
         self.d_max_ = None
         self.is_fitted_ = False
+        self.labels_ = np.zeros((0,), dtype=int)
 
     def __getattr__(self, key):
         if key in self.params:
