### 2025-07-01 #7764 @mstimberg
@chenxinye and @TahiriNadia thank you for your comments/suggestions.

@chenxinye Are you happy with the reply in https://github.com/openjournals/joss-reviews/issues/7764#issuecomment-2899216018 or do you recommend some additional work to be done for the paper?

@TahiriNadia Did you find time to have a closer look at the software/paper with respect to the remaining checklist items? And do you find the reply in https://github.com/openjournals/joss-reviews/issues/7764#issuecomment-2899202094 satisfactory? A minor remark from my side with respect to "Whether the software is sufficiently useful that it is likely to be cited by your peer group."  Note that JOSS does not strictly require that the software has already been cited – we only require that it is *likely to be cited* in the future, which in particular means that it should be useful for other people than its author(s).
