### 2025-07-10 #7764 @chenxinye
I am happy with it, good software! NO further comments now.

Also, I will be very interested if it is possible to integrate CLASSIX (https://github.com/nla-group/classix) in your software.

@NiklasMelton
