### 2025-07-13 #7764 @chenxinye
@NiklasMelton, CLASSIX is an density-based clustering algorithm, offering remarkable speed and performance compared to DBSCAN. I believe your first point is a promising direction, and I’d love to collaborate on a short paper to explore this further!

I would need to learn more on ART. Let’s continue this discussion via email ([xinyechenai@gmail.com](mailto:xinyechenai@gmail.com)). Looking forward to staying in touch!

P.S. It was truly a privilege to review your paper!
