### 2025-07-14 6763b09
fix bug with fusionART weight updates on certain art modules
