### 2025-07-14 7ec7cee
fix bug with fusionART weight updates on certain art modules
