### 2025-07-14 e707ae0
fix bug with fusionART weight updates on certain art modules
