### 2025-07-14 b123eb1
fix bug with fusionART weight updates on certain art modules
