### 2025-07-14 #7764 @NiklasMelton
@chenxinye I think this could make for a good short paper. You can email me at `niklasmelton@mst.edu`.

The simple answer is that this hybrid would be either a batch model or (at best) a mini-batch model. There do exist incremental PCA algorithms but that doesn't help the sorting if we are still only getting one sample at a time. But storing samples in a limited memory cache might allow it to become pseudo-incremental. I would also recommend you look into VAT:

Bezdek, J. C., & Hathaway, R. J. (2002).
VAT: A tool for visual assessment of cluster tendency.
Proceedings of the 2002 International Joint Conference on Neural Networks.
doi:10.1109/IJCNN.2002.1007487

This is a rather interesting algorithm developed for visualizing clusters but has been shown to be very effective at providing an unsupervised presentation ordering. It might be worth exploring it for CLASSIX as well. I actually have an implementation in the artily repo here: https://github.com/NiklasMelton/AdaptiveResonanceLib/blob/develop/artlib/common/VAT.py

Looking forward to keeping in touch as well. Thank you for your time as a reviewer!
