### 2025-07-16 fb94b03
plt.scatter -> ax.scatter
