### 2025-07-16 92292e3
same fix for spherical alpha art
