### 2025-07-19 NiklasMelton/AdaptiveResonanceLib#174 @TahiriNadia
[JOSS] 💡 Workflow Design Diagram for AdaptiveResonanceLib
Dear,

I would like to suggest to create a lightweight architectural diagram or flowchart that shows how the main components of `AdaptiveResonanceLib` (e.g., input formats, ART variants, training loops, evaluation modules) interact.

🧠 The goal is not to repeat what is in the papers, but to offer a system-level overview that helps users and contributors understand the modularity and extensibility of the library — similar to what you might find in a “System Overview” section of a well-documented ML toolkit.
