+++ CITATION.cff
@@ -1,7 +1,7 @@
 cff-version: 1.2.0
 message: "If you use this software, please cite it as below."
 title: "AdaptiveResonanceLib"
-version: "0.1.5"
+version: "0.1.6"
 doi: "10.5281/zenodo.9999999"
 authors:
   - family-names: "Melton"
