### 2025-10-17 #7764 @crvernon
:wave: @NiklasMelton - I just need you to address the following before I accept this for publication:

- [ ] LINE 74:  The "p" in "python" should be capitalized.
- [ ] LINE 95:  I believe "art" here should also be capitalized

Let me know when these changes have been made and I'll accept.  Thank you.
